Set all descriptions, not only the English ones
This is a temporary workaround to include all descriptions in the spot description. In previous versions we were only retrieving English descriptions and ignoring the others. Due to that, when Germans were loading a spot in Germany for example, sometimes they would get no description for that spot because the descriptions were saved in German language.